Rename `commitment_databases` module to `databases`
